+++ src/components/site/DatePicker.jsx
@@ -9,6 +9,7 @@ import {
 import { ChevronDownIcon } from "lucide-react";
 import { useState } from "react";
 import MotionButton from "@components/application/MotionButton.jsx";
+import { Button } from "@components/ui/button.jsx";
 
 const DatePicker = ({ value, onChange, ...props }) => {
   const [open, setOpen] = useState(false);
@@ -17,16 +18,19 @@ const DatePicker = ({ value, onChange, ...props }) => {
     <Container>
       <Popover open={open} onOpenChange={setOpen}>
         <PopoverTrigger asChild>
-          <MotionButton
+          <Button
             id="date"
+            variant="outline"
             className={cn(
-              "hover:bg-accent hover:text-accent-foreground/ dark:hover:bg-accent/50 focus:ring-ring focus:ring-offset-background flex w-full items-center justify-between rounded-md border px-3 py-2 font-normal focus:ring-2 focus:ring-offset-2 focus:outline-none",
-              { "text-foreground": value, "text-foreground/50": !value },
+              "text-foreground/60 w-full justify-between font-normal",
+              {
+                "text-foreground": value,
+              },
             )}
           >
             {value ? value.toLocaleDateString() : "Select date"}
             <ChevronDownIcon size={15} />
-          </MotionButton>
+          </Button>
         </PopoverTrigger>
         <PopoverContent className="w-auto overflow-hidden p-0" align="start">
           <Calendar