+++ src/dev-data/finance/card.data.js
@@ -1,0 +1,28 @@
+import { BiDollar, BiTrendingUp } from "react-icons/bi";
+import { GrDocument } from "react-icons/gr";
+
+const cardData = [
+  {
+    id: 1,
+    title: "Total Revenue",
+    icon: BiDollar,
+    iconColor: "#00c951",
+    number: "650",
+  },
+  {
+    id: 2,
+    title: "Completed Visits",
+    icon: GrDocument,
+    iconColor: "#007bff",
+    number: "2",
+  },
+  {
+    id: 3,
+    title: "Average Visit Value",
+    icon: BiTrendingUp,
+    iconColor: "#8e44ad",
+    number: "325",
+  },
+];
+
+export { cardData };